Remove data encryption code and documentation
