--- Infrastructure/Configuration/DataEncryption.cs
@@ -1,7 +1,0 @@
-﻿namespace Infrastructure.Configuration;
-
-public class DataEncryption
-{
-    public string SourceDirectory { get; set; } = string.Empty;
-    public string Key { get; set; } = string.Empty;
-}
